--- apps/charterafrica/src/components/DocumentList/index.js
@@ -1,3 +1,0 @@
-import DocumentList from "./DocumentList";
-
-export default DocumentList;
