+++ apps/charterafrica/src/components/Documents/index.js
@@ -1,0 +1,3 @@
+import Documents from "./Documents";
+
+export default Documents;
